дз

+++ src/Main.java
@@ -2,7 +2,7 @@ import java.util.Arrays;
 
 public class Main {
 
-    public static void quickSort(int[] array, int low, int high) {
+    public static void quickSort(long[] array, long low, long high) {
         if (array.length == 0)
             return;
 
@@ -10,24 +10,24 @@ public class Main {
             return;
 
 
-        int middle = low + (high - low) / 2;
-        int opora = array[middle];
+        long middle = low + (high - low) / 2;
+        long opora = array[(int) middle];
 
 
-        int i = low, j = high;
+        long i = low, j = high;
         while (i <= j) {
-            while (array[i] < opora) {
+            while (array[(int) i] < opora) {
                 i++;
             }
 
-            while (array[j] > opora) {
+            while (array[(int) j] > opora) {
                 j--;
             }
 
             if (i <= j) {
-                int temp = array[i];
-                array[i] = array[j];
-                array[j] = temp;
+                long temp = array[(int)i];
+                array[(int)i] = array[(int)j];
+                array[(int)j] = temp;
                 i++;
                 j--;
             }
@@ -40,12 +40,12 @@ public class Main {
             quickSort(array, i, high);
     }
     public static void main(String[] args) {
-        int[] x = { 15, -7, 10, 7, 3, -2, 0, 12, 6 };
+        long[] x = { 15, -7, 10, 7, 3, -2, 0, 12, 6 };
         System.out.println("Было");
         System.out.println(Arrays.toString(x));
 
-        int low = 0;
-        int high = x.length - 1;
+        long low = 0;
+        long high = x.length - 1;
 
         quickSort(x, low, high);
         System.out.println("Стало");
